Player behavior improvements (#2338)
* Player behavior improvements

Replaces Earth Tux’s ability with a controllable rolling rock
Glide time for Air Tux is infinite to stay in line with Earth Tux
Made Air Tux’s trampoline bounce less crazy
Made ghost mode a bit more like swimming (move using dir keys, boost using jump)
Remove unneeded “m_shooting_timer” variable

* Some updates [read desc]

More proper hitbox, decreased jump height while stone, harder controls when stone, only turn to stone when going faster than crawling, separated stone variables for easier access, fixed stone acceleration variable errors

* Fix an error

* More updates [read description]

Lower max roll speed, new function for ending rolling, catch-all for improper hitbox adjustments, stop being stone when hitting water, only start rolling on-ground, fix weird skid hitbox issue

* More fixes [read description]

- Separated swimming variables
- Fixed bug where swimboosting would override external velocity, such as by bumpers, trampolines or wind
- Tux breaks out of rock mode when hitting a wall at high speed
- Added a parameter for if Tux is being "violently" knocked out of rock mode

* Bumper makes a little more sense underwater

* Add particles to breaking out of rock mode

* Fix conversion error

* Stronger buttjump (test)

If unliked, it can be removed

* Fix merge-conflict-based errors

* Incorporate Rusty's suggestions

+++ src/object/explosion.cpp
@@ -113,7 +113,7 @@ Explosion::explode()
       Vector add_speed = glm::normalize(direction) * force;
 
       auto player = dynamic_cast<Player*>(obj);
-      if (player) {
+      if (player && !player->is_stone()) {
         player->add_velocity(add_speed);
       }
 
@@ -176,7 +176,7 @@ Explosion::collision(GameObject& other, const CollisionHit& )
     return ABORT_MOVE;
 
   auto player = dynamic_cast<Player*>(&other);
-  if (player != nullptr) {
+  if (player != nullptr && !player->is_stone()) {
     player->kill(false);
   }
